pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
tau2-evaluator = [
    "tau2 @ git+https://github.com/sierra-research/tau2-bench.git",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
